requirements.txt
pandas
pytorch-cpu
torchvision-cpu